wrote test-cases for services

+++ src/main/java/com/foodorderingapp/serviceImpl/FoodServiceImpl.java
@@ -1,9 +1,12 @@
 package com.foodorderingapp.serviceImpl;
 
+import com.foodorderingapp.commons.PageModel;
 import com.foodorderingapp.dao.FoodDAO;
 import com.foodorderingapp.dao.RestaurantDAO;
+import com.foodorderingapp.exception.DataNotFoundException;
 import com.foodorderingapp.model.Food;
 import com.foodorderingapp.service.FoodService;
+import jdk.nashorn.internal.runtime.ECMAException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,35 +27,78 @@ public class FoodServiceImpl implements FoodService{
     }
 
     public boolean deleteFood(Food food) {
-        return foodDAO.deleteFood(food);
+        if (food==null) {
+            throw new DataNotFoundException("cannot delete food.");
+        } else {
+            return foodDAO.deleteFood(food);
+        }
     }
 
     public boolean updateFood(Food food,int id) {
         Food food1 = foodDAO.getFoodById(id);
+        if(food1==null){
+            throw new DataNotFoundException("cannot find food.");
+        }
         food1.setName(food.getName());
         food1.setPrice(food.getPrice());
         return foodDAO.updateFood(food1);
     }
 
     public List<Food> getAll() {
-        return foodDAO.getAll();
+        List<Food> foodList= foodDAO.getAll();
+        if(foodList==null ){
+            throw new DataNotFoundException("cannot find foodlist.");
+        }else{
+            return foodList;
+        }
     }
 
     public Food getFoodById(int id) {
-        return foodDAO.getFoodById(id);
+        Food food= foodDAO.getFoodById(id);
+        if(food==null){
+            throw new DataNotFoundException("cannot find food.");
+        }else{
+            return food;
+        }
     }
 
     public List<Food> getFoodByRestaurantId(int id) {
-        return foodDAO.getFoodByRestaurantId(id);
+        List<Food> foodList=foodDAO.getFoodByRestaurantId(id);
+        if(foodList==null || foodList.size()==0){
+            throw new DataNotFoundException("cannot find foodList.");
+        }
+        else{
+            return foodList;
+        }
     }
 
     public List<Food> addFoodsToRestaurant(List<Food> foodList) {
-        return foodDAO.addFoodsToRestaurant(foodList);
+
+        if (foodList == null || foodList.size()==0) {
+            throw new DataNotFoundException("cannot add foodList.");
+        } else {
+            return foodDAO.addFoodsToRestaurant(foodList);
+        }
     }
 
     @Override
     public Food getFoodByResName(String restaurantName, String foodName) {
-        return foodDAO.getFoodByResName(restaurantName,foodName);
+
+        Food food=foodDAO.getFoodByResName(restaurantName,foodName);
+        if(food==null){
+            throw new DataNotFoundException("cannot find foodList1.");
+        }else{
+            return food;
+        }
     }
 
+    @Override
+    public List<Food> getPaginatedFood(PageModel pageModel, int id) {
+        return foodDAO.getPaginatedFood(pageModel,id);
+    }
+
+    @Override
+    public long countFood(int id) {
+        return foodDAO.countFood(id);
+    }
 }
